requirements.txt
Flask
google-generativeai
PyPDF2
requests
dotenv